Move hardware components to specific package
