+++ src/main/java/com/team1816/lib/hardware/CtreMotorFactory.java
@@ -4,6 +4,8 @@ import com.ctre.phoenix.ErrorCode;
 import com.ctre.phoenix.ParamEnum;
 import com.ctre.phoenix.motorcontrol.*;
 import com.ctre.phoenix.motorcontrol.can.*;
+import com.team1816.lib.hardware.components.GhostTalonSRX;
+import com.team1816.lib.hardware.components.LazyTalonFX;
 import com.team254.lib.drivers.LazyTalonSRX;
 
 /**
