--- src/main/java/com/team1816/lib/hardware/GhostSolenoidBase.java
@@ -1,5 +1,0 @@
-package com.team1816.lib.hardware;
-
-public interface GhostSolenoidBase {
-
-}
