+++ src/main/java/com/team1816/lib/hardware/components/GhostSolenoidBase.java
@@ -1,0 +1,5 @@
+package com.team1816.lib.hardware.components;
+
+public interface GhostSolenoidBase {
+
+}
