+++ src/main/java/com/team1816/lib/hardware/components/GhostTalonSRX.java
@@ -1,4 +1,4 @@
-package com.team1816.lib.hardware;
+package com.team1816.lib.hardware.components;
 
 import com.ctre.phoenix.ErrorCode;
 import com.ctre.phoenix.ParamEnum;
