+++ src/main/java/com/team1816/lib/hardware/components/ILazyMotorControllerEnhanced.java
@@ -1,4 +1,4 @@
-package com.team1816.lib.hardware;
+package com.team1816.lib.hardware.components;
 
 import com.ctre.phoenix.motorcontrol.IMotorControllerEnhanced;
