+++ src/main/java/com/team1816/lib/hardware/components/LazyTalonFX.java
@@ -1,9 +1,9 @@
-package com.team1816.lib.hardware;
+package com.team1816.lib.hardware.components;
 
 import com.ctre.phoenix.motorcontrol.ControlMode;
 import com.ctre.phoenix.motorcontrol.can.TalonFX;
 
-public class LazyTalonFX extends TalonFX implements ILazyMotorControllerEnhanced  {
+public class LazyTalonFX extends TalonFX implements ILazyMotorControllerEnhanced {
     protected double mLastSet = Double.NaN;
     protected ControlMode mLastControlMode = null;
